Reorganise nodewallet cmd file to better scope new commands
