+++ cmd/vega/nodewallet/import.go
@@ -1,0 +1,83 @@
+package nodewallet
+
+import (
+	"fmt"
+
+	"code.vegaprotocol.io/vega/config"
+	"code.vegaprotocol.io/vega/fsutil"
+	"code.vegaprotocol.io/vega/logging"
+	"code.vegaprotocol.io/vega/nodewallet"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/jessevdk/go-flags"
+)
+
+type importCmd struct {
+	Config nodewallet.Config
+
+	WalletPassphrase config.Passphrase `short:"w" long:"wallet-passphrase"`
+
+	Chain      string `short:"c" long:"chain" required:"true" description:"The chain to be imported (vega, ethereum)"`
+	WalletPath string `long:"wallet-path" required:"true" description:"The path to the wallet file to import"`
+	Force      bool   `long:"force" description:"Should the command re-write an existing nodewallet file if it exists"`
+	Help       bool   `short:"h" long:"help" description:"Show this help message"`
+}
+
+func (opts *importCmd) Execute(_ []string) error {
+	if opts.Help {
+		return &flags.Error{
+			Type:    flags.ErrHelp,
+			Message: "vega nodewallet import subcommand help",
+		}
+	}
+	log := logging.NewLoggerFromConfig(logging.NewDefaultConfig())
+	defer log.AtExit()
+
+	if ok, err := fsutil.PathExists(rootCmd.RootPath); !ok {
+		return fmt.Errorf("invalid root directory path: %w", err)
+	}
+
+	nodePass, err := rootCmd.PassphraseFile.Get("node wallet")
+	if err != nil {
+		return err
+	}
+
+	walletPass, err := opts.WalletPassphrase.Get("blockchain wallet")
+	if err != nil {
+		return err
+	}
+
+	conf, err := config.Read(rootCmd.RootPath)
+	if err != nil {
+		return err
+	}
+	opts.Config = conf.NodeWallet
+
+	if _, err := flags.NewParser(opts, flags.Default|flags.IgnoreUnknown).Parse(); err != nil {
+		return err
+	}
+
+	ethClient, err := ethclient.Dial(opts.Config.ETH.Address)
+	if err != nil {
+		return err
+	}
+
+	nw, err := nodewallet.New(log, conf.NodeWallet, nodePass, ethClient, rootCmd.RootPath)
+	if err != nil {
+		return err
+	}
+
+	_, ok := nw.Get(nodewallet.Blockchain(opts.Chain))
+	if ok && opts.Force {
+		log.Warn("a wallet is already imported for the current chain, this action will rewrite the import", logging.String("chain", opts.Chain))
+	} else if ok {
+		return fmt.Errorf("a wallet is already imported for the chain %v, please rerun with option --force to overwrite it", opts.Chain)
+	}
+
+	err = nw.Import(opts.Chain, nodePass, walletPass, opts.WalletPath)
+	if err != nil {
+		return err
+	}
+
+	fmt.Println("import success")
+	return nil
+}
